Add files via upload
The above is the code for my first Project under the CantiLever Internship. 
Project Title: Contact Book
Here i have implemented a simple code to create a contact book that saves, deletes, views, edits resets and exits the page. I have implemented it using GUI (Tkinter library in python) to make it more Aesthetically pleasing as given in the requirment